fix sharing on iPads
The YouTube app calls entirely different methods when sharing something
on an iPad instead of an iPhone. This commit hooks additional methods
that are used on iPads instead. It's a bit hacky, but it works. Since
I don't have an iPad myself, I developed and tested this using the
iPad mode of the YTLitePlus app: https://github.com/YTLitePlus/YTLitePlus

Fix #2