fix sharing on iPads 2.0
Hook `YTAccountScopedCommandResponderEvent` instead of causing a null
pointer exception by returning NULL on `YTShareRequestViewController`
init. The disadvantage of this is, that it's not only called when
sharing something, but also for other events, e.g. a few times when
starting to play a video. However, since the hook returns early and it's
not like called every second, it shouldn't be problematic.

Fix #2 again